Set up utop to work with nix nicely
The nix expression for utop did not set environment variables correctly,
so I made my nix expression depend on the unstable channel.

Making the ocaml toplevel work with janestreet core also took a little
bit of effort. The current solution is to pass a $findlib variable from
the nix expression. This could be improved upon, but it's pretty simple
as it is.

Add a couple make targets; the makefile shouldn't ever do anything
except be a dumping ground for common sequences of commands.