Change almost all collections to sequences.
Almost everything is incremental now. This is possible because we are no
longer splitting on common letters. Instead of ordering based on number
of candidates, we order based on number of letters instead. A possible
further improvement is to order based on new letters or perhaps based on
size of effect.