Convert Translation to using association lists.
Instead of using two arrays of length 26 to store translations, use a
single association list. The problem is that although the arrays have
very fast indexing, we must copy two of them every time we want to
overlay a translation, we must copy both arrays. With association lists,
we only need to concatenate the lists. Indexing is much slower, but we
index about as much as we overlay, so it shouldn't be a problem.
Implementation is much simpler as well.